requirements.txt
-i https://pypi.org/simple
cfgv==3.4.0; python_version >= '3.8'
distlib==0.3.9
factory-boy==3.3.1; python_version >= '3.8'
faker==33.1.0; python_version >= '3.8'
filelock==3.16.1; python_version >= '3.8'
identify==2.6.3; python_version >= '3.9'
iniconfig==2.0.0; python_version >= '3.7'
nodeenv==1.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
packaging==24.2; python_version >= '3.8'
platformdirs==4.3.6; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
pre-commit==4.0.1; python_version >= '3.9'
pytest==8.3.4; python_version >= '3.8'
pytest-django==4.9.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.12.2; python_version >= '3.8'
virtualenv==20.28.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
django==5.1.3; python_version >= '3.10'
djangorestframework==3.15.2; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
sqlparse==0.5.2; python_version >= '3.8'
